change readme and code

+++ fib/rust/src/lib.rs
@@ -1,7 +1,7 @@
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
-pub fn fib(n: i64) -> i64 {
+pub fn fib(n: i32) -> i32 {
     if n <= 1 {
         return n
     }
